backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.40
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

